fix: remove explicit any from Giftcard
