+++ packages/lib/src/components/internal/SecuredFields/lib/types.ts
@@ -351,3 +351,12 @@ export interface SFKeyDownObj {
     fieldType: string;
     action: string;
 }
+
+export interface BrandConfiguration {
+    name?: string;
+    icon?: string;
+}
+
+export interface CardBrandsConfiguration {
+    [key: string]: BrandConfiguration;
+}
